# Improve Query performance and return an error if the query